panel index completed.

+++ bakimonarim.entity/VGrup.cs
@@ -6,8 +6,8 @@ namespace bakimonarim.entity
     public class VGrup : IEntity
     {
         [Key]
-        public int GrupID { get; set; }
-        public int AracSinifID { get; set; }
+        public long GrupID { get; set; }
+        public long AracSinifID { get; set; }
         public string GrupMarka { get; set; }
         public string GrupModel { get; set; }
         public string LastikTip { get; set; }
